implemented remaining functions

+++ src/main/java/entity/Prescription.java
@@ -1,7 +1,8 @@
 package entity;
 
 /**
- * Represents a medication prescribed to a patient in the Hospital Management System.
+ * Represents a medication prescribed to a patient in the Hospital Management
+ * System.
  */
 public class Prescription {
 
@@ -11,7 +12,7 @@ public class Prescription {
     public enum PrescriptionStatus {
         PENDING, DISPENSED
     }
-    
+
     /**
      * Unique identifier for the prescription.
      */
@@ -21,22 +22,22 @@ public class Prescription {
      * Unique identifier for the appointment.
      */
     private String appointmentID;
-    
+
     /**
      * Unique identifier for the medication prescribed.
      */
     private String medicationID;
-    
+
     /**
      * Quantity of the medication prescribed.
      */
-    private int quantity;
+    private double quantity;
 
     /**
      * The current status of the prescription.
      */
     private PrescriptionStatus status;
-    
+
     /**
      * Notes for the prescription.
      */
@@ -44,6 +45,7 @@ public class Prescription {
 
     /**
      * Constructor for Prescription.
+     *
      * @param prescriptionID The unique identifier for the prescription
      * @param appointmentID The unique identifier for the appointment
      * @param medicationID The unique identifier for the medication prescribed
@@ -51,7 +53,7 @@ public class Prescription {
      * @param status The current status of the prescription
      * @param notes The notes for the prescription
      */
-    public Prescription(String prescriptionID, String appointmentID, String medicationID, int quantity, PrescriptionStatus status, String notes) {
+    public Prescription(String prescriptionID, String appointmentID, String medicationID, double quantity, PrescriptionStatus status, String notes) {
         this.prescriptionID = prescriptionID;
         this.appointmentID = appointmentID;
         this.medicationID = medicationID;
@@ -59,17 +61,19 @@ public class Prescription {
         this.status = status;
         this.notes = notes;
     }
-    
+
     /**
      * Get the unique identifier for the prescription.
+     *
      * @return The prescription ID
      */
     public String getPrescriptionID() {
         return prescriptionID;
     }
-    
+
     /**
      * Get the unique identifier for the appointment.
+     *
      * @return The appointment ID
      */
     public String getAppointmentID() {
@@ -78,6 +82,7 @@ public class Prescription {
 
     /**
      * Get the unique identifier for the medication prescribed.
+     *
      * @return The medication ID
      */
     public String getMedicationID() {
@@ -86,14 +91,16 @@ public class Prescription {
 
     /**
      * Get the quantity of the medication prescribed.
+     *
      * @return The quantity
      */
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
     /**
      * Get the current status of the prescription.
+     *
      * @return The prescription status
      */
     public PrescriptionStatus getStatus() {
@@ -102,14 +109,16 @@ public class Prescription {
 
     /**
      * Get the notes for the prescription.
+     *
      * @return The notes
      */
     public String getNotes() {
         return notes;
     }
-    
+
     /**
      * Set the unique identifier for the prescription.
+     *
      * @param prescriptionID The prescription ID to set
      */
     public void setPrescriptionID(String prescriptionID) {
@@ -118,6 +127,7 @@ public class Prescription {
 
     /**
      * Set the unique identifier for the appointment.
+     *
      * @param appointmentID The appointment ID to set
      */
     public void setAppointmentID(String appointmentID) {
@@ -126,6 +136,7 @@ public class Prescription {
 
     /**
      * Set the unique identifier for the medication prescribed.
+     *
      * @param medicationID The medication ID to set
      */
     public void setMedicationID(String medicationID) {
@@ -134,14 +145,16 @@ public class Prescription {
 
     /**
      * Set the quantity of the medication prescribed.
+     *
      * @param quantity The quantity to set
      */
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 
     /**
      * Set the current status of the prescription.
+     *
      * @param status The prescription status to set
      */
     public void setStatus(PrescriptionStatus status) {
@@ -150,12 +163,13 @@ public class Prescription {
 
     /**
      * Set the notes for the prescription.
+     *
      * @param notes The notes to set
      */
     public void setNotes(String notes) {
         this.notes = notes;
     }
-    
+
     @Override
     public String toString() {
         return "Prescription{" + prescriptionID + "," + appointmentID + "," + medicationID + "," + quantity + "," + status + "," + notes + "}";
